requirements.txt
numpy
numpy-quaternion
torch
scipy
pillow
pyopencl
Mako
plyfile